requirements.txt
flask
flask_cors
whisper
torch
deep-translator
opencv-python
srt
numpy
pillow
gunicorn
